Improve layout of Public Body CSV Import
Moves the help section below the form to make the actual input options clearer.

Move the "Standard tags" to a sidebar next to the help as the list can
get pretty long (see WDTK).

Minor rewording of the link to alaveteli.org documentation to be more
inline with Admin Censor Rules inline help.